Create SourceCode.gs
The user will input the list of meeting names and time onto the Google Spreadsheet. The code will get all meetings from every Pro accounts and sort it by time ascending. Based on the given date and time that the user wants to create a meeting, it will find the Zoom account that is available 30 minutes prior to the meeting.

*Note: If there are multiple meetings that occurs at the same time, the code will assign different account to the simultaneous meetings.